WIP: Package cleanup
