+++ src/lib/kotlin/slatekit-apis/src/main/kotlin/slatekit/apis/support/CallSupport.kt
@@ -9,7 +9,7 @@ import slatekit.common.ext.toResponse
 import slatekit.results.Outcome
 import slatekit.results.Try
 
-interface ExecSupport {
+interface CallSupport {
 
     fun host(): ApiServer
