+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/apis/Api_Naming_Tests.kt
@@ -5,7 +5,7 @@ import org.junit.Assert
 import org.junit.Test
 import slatekit.apis.routes.Api
 import slatekit.apis.ApiServer
-import slatekit.apis.ApiSettings
+import slatekit.apis.Settings
 import slatekit.apis.Verb
 import slatekit.common.naming.LowerHyphenNamer
 import slatekit.common.naming.LowerUnderscoreNamer
@@ -19,7 +19,7 @@ class Api_Naming_Tests : ApiTestsBase() {
 
     @Test fun can_use_naming_convention_lowerHyphen() {
         val apis = ApiServer(ctx, apis = listOf(Api(SamplePOKOApi::class,
-                "app", "SamplePOKO")), settings = ApiSettings(naming = LowerHyphenNamer())
+                "app", "SamplePOKO")), settings = Settings(naming = LowerHyphenNamer())
         )
         Assert.assertTrue( apis.get("app"   , "sample-poko", "get-time"    ).success)
         Assert.assertTrue(!apis.get("app"   , "SamplePOKO" , "getTime"     ).success)
@@ -41,7 +41,7 @@ class Api_Naming_Tests : ApiTestsBase() {
     @Test fun can_use_naming_convention_lowerUnderscore() {
         val apis = ApiServer(ctx, apis = listOf(Api(SampleExtendedApi::class,
                 "app", "SampleExtended", declaredOnly = false)),
-                settings = ApiSettings(naming = LowerUnderscoreNamer())
+                settings = Settings(naming = LowerUnderscoreNamer())
         )
         Assert.assertTrue( apis.get("app"   , "sample_extended", "get_seconds" ).success)
         Assert.assertTrue( apis.get("app"   , "sample_extended", "get_time"    ).success)
